Allow observers to return a state
Allow Observers to return a state, if the state is not equal to the
status the Observer was called with, the Informer will re-inform all the
Observers of the new state.

Rename ObserverFuncs to Factory, this better represents the job that the
func does. Add an ObserverFactory interface and update the funcs to be
adaptor functions that implement the interface.

Also:
- remove the Informer.Inform method, it leaked the ID of the Observers
  that were registered with the Informer, and there was confusing
  overlap between the Exec ID and the Observer ID leading to bugs.
- Add tests for the polling informer managers / informers / observers.
- Add an internal fakes directory for the Driver / Observer / Store,
  to make testing easier.
- Use go generate ./... to re-generate the fakes.
- Add Makefile, testing, linting
- Add github actions