Add an AWS provider (#37)
* Add an AWS provider

AWS Provider
============

Add a new package `providers/awsprov`. This package contains the aws
provider driver. Implement the driver for AWS.

The AWS driver starts a new t2.mico instance. It assumes the default
VPC, and subnets. The driver makes use of the UserData script, which is
installed on the machine and executed on first boot.
This script
- registers the public keys for SSH access,
- creates an `unweave` user for use on the instance
- attaches, mounts, and formats any EBS volumes specified.

The driver attaches tags for the ExecID, ProjectID, and UserID to the
AWS instance.

The AWS provider implements volumes using EBS.

In order to mount the EBS volumes, the EC2 instance needs an
InstanceProfile with the correct permissions. The AWS Driver will look
for the InstanceProfile called `UnweaveEc2ExecInstanceProfile`. If this
profile does not exist, it will be created and have the correct
permissions attached to it. This allows the UserData script to assume
the correct permissions in order to mount the required EBS volumes.

The state of an instance depends on two fields:
- InstanceState
- InstanceSummaryStatus

The InstanceState can be running, but the instance might still not be
contactable. We wait until the 'healthchecks' have passed and
InstanceSummaryStatus is marked as 'ok' before marking the Instance (and
Exec) as running and attempting to SSH to it.

Exec Service
============

The Exec Driver has a new method called `ExecConnectionInfo` this is
added because the service (and the driver the called is delegated to) is
mostly likely to be able to fetch the ConnectionInfo. Generally IP
addresses are given to Execs after they have started, so we have to
consult the underlying Service/Driver to query for this information.

Once an Exec is "running" we can call the Service's RefreshConnetionInfo
to get the new connection info and persist it.

Default Exec images
==================

There are now two fields for the DefaultExecImage. The
GlobalDefaultImageURI, and a Service specific ServiceDefaultImage. This
allows the AWS Exec Service to have it's own default image (an amazon
machine image, or 'ami') that is different from the
GlobalDefaultImageURI which is a docker image URI.

Observers
========

The standard polling state observer is now uses the Service to refresh
the connection info when an Exec transitions to the Running state. It
will also record that Exec as running in the database.

Currently not supported
=======================

1. Any other instance type except for t2.micro.
2. Exposing an HTTP service on an AWS exec.
3. VPC, subnet, or security group config.

Handling providers
==================

Each of the main services (execsrv, volumesrv, providersrv) each have a
Router (i.e. ExecRouter, VolumeRouter, etc). The router is built with
one Service for each provider, and the router handles delegating to the
correct Service based on the provider needed.

* Add 10m timeout to golangci-lint

* Fix deadlock in async tests

* Add GPU nodes and node calculation from HardwareSpec

* Add Provider method to ProviderDriver

* Return structured error on HTTPService