Endpoints and Evals (#40)
* Endpoints and Evals

Add endpoints and evals, add db tables, models, handlers and services.

Endpoints:
- an abstraction over an exec, that exposes it on an HTTP address.

Eval:
- an abstraction over an exec, where the exec should expose two urls:
  /dataset and /assert. These urls are used to get the input data to
  test and endpoint with, and assert the success or failure of a single
  item in the dataset.

Also:
- move service delegation out of routers and into services, so that
  other services can make use of the ExecService delegation (based on
  provider).

* Fix linters and review changes

* Update http errors and add FK constraints on db

* Update modules

* Endpoint checks

Add endpoint checks. Checks are a wrapper around all the evals that need
to be run against a specific endpoint. Checks track the progress of
those evals.

* Add endpoint status and created at time

* Endpoints and evals carry and http address

This means we don't need to fetch the exec in order to get the http
address, and we can name the endpoint after the endpoint / eval.

* Fix linter errors