feat: Extensible eval checks
Allow evals to serve an eval manifest in `/` where the user can pick
to customise which endpoints to call for dataset, assert, and running
the model.

By default running the model calls the endpoint directly. However, you
can call another endpoint within the eval by returning a path in the
eval manifest's runURL.

As an example, `"runURL": "/run"` would run the model against
`<eval-url>/run` instead of `<endpoint-url>`.

The model run receives a header `X-Unweave-Target-Endpoint-URL` so the
eval knows which endpoint we are targetting in the check.

This allows optional extensibility for users who need to do some data
processing to call the model based from the dataset input.

You just need to add:

```python
@app.get("/")
async def manifest():
  return {"runURL": "/run"}

@app.post("/run")
ascync def run_check([...]):
  # call model URL from `X-Unweave-Target-Endpoint-URL`
```

This PR also includes a small code re-organisation, and better error
handling of calls to `/dataset` and `/assert`.